% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myfacto.R
\name{my_facto_r}
\alias{my_facto_r}
\title{Wrapper to fortran version of computing factorials}
\usage{
my_facto_r(num)
}
\arguments{
\item{num}{number for which factorial should be computed}
}
\value{
retval$answer resulting value for factorial
}
\description{
We use a fortran subroutine to compute factorials.
This compiled subroutine is provided in the libs
directory of this package. This function provides
an R-Interface to this compiled subroutine.
}
